requirements.txt
parsel==1.6.0
pywebview==3.6.3
PyYAML==6.0
rarfile==4.0
requests==2.28.1